Deleted ProgrammingChannel cog, moved to misc
